Move `precompile` -> `lib/precompile`
