+++ src/lib/do-transform.js
@@ -1,11 +1,10 @@
-import {configureEnvironment} from '../precompile/configure-environment';
-import {getUsedDependencies} from '../precompile/get-used-dependencies';
-import {loadDependencies} from '../precompile/load-dependencies';
-import {precompileToLocalVar} from '../precompile/precompile-to-local-var';
-
 import {hasAsyncTags} from './ast/has-async-tags';
 import {getLoaderOutput} from './output/get-loader-output';
 import {getTemplateImports} from './output/get-template-imports';
+import {configureEnvironment} from './precompile/configure-environment';
+import {getUsedDependencies} from './precompile/get-used-dependencies';
+import {loadDependencies} from './precompile/load-dependencies';
+import {precompileToLocalVar} from './precompile/precompile-to-local-var';
 
 
 const staticExtensionPath = require.resolve(
