+++ src/lib/precompile/configure-environment.js
@@ -1,6 +1,6 @@
 import nunjucks from 'nunjucks';
 
-import {getModule} from '../public/utils/get-module';
+import {getModule} from '../../public/utils/get-module';
 
 /**
  * @param {Object}   options
