+++ src/lib/precompile/precompile-to-local-var.js
